Issue default Content-Length wherever possible
There are several scenarios where we are serving content without setting
the Content-Length header. Today have a safeguard in place to force the
connection close in the absence of a Content-Length, since HTTP 1.1 uses
keepalive connections by default. However in many of these cases we know or
can easily determine the content length ahead of time, so we may as well
be consistent about it.

Note that the default Content-Length will not overwrite any explicitly-
specified Content-Length header (set by pipes, .header files, Python etc.)
This keeps in the spirit of wptserve being a well-behaved HTTP server by
default, but which can be customized even to send invalid HTTP responses
in order to test how browsers handle them.